+++ magick/gem.h
@@ -21,8 +21,8 @@ extern MagickExport int
   GetOptimalKernelWidth2D(const double,const double);
 
 extern MagickExport PixelPacket
-  AlphaComposite(const PixelPacket *,const unsigned long,const PixelPacket *,
-    const unsigned long),
+  AlphaComposite(const PixelPacket *,const double,const PixelPacket *,
+    const double),
   InterpolateColor(const Image *,const double,const double,ExceptionInfo *);
 
 extern MagickExport Quantum
